stage3

+++ dop/src/Cat.kt
@@ -1,2 +1,2 @@
-data class Cat(val color: String, val age: Int, val weight: Int) {
+data class Cat(val color: String, val age: Int, val weight: Float) {
 }
